move try_spinner & run_shell_commad
