+++ src/capi/azext_capi/helpers/spinner.py
